Fix Windows Build Warnings
This PR resolves Windows build warnings.

- Type Conversion Warnings (C4244, C4267, C4146, C4293)
- MSVC Compiler Flag Warnings (D9024, D9025, D9027)
- Hexadecimal Floating-Point Warnings (C4305, C4309)

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped

+++ nntrainer/dataset/databuffer.cpp
@@ -197,7 +197,7 @@ void DataBuffer::displayProgress(const int count, float loss) {
     else
       progress = (((float)(count * batch_size)) / (float)samples_per_epoch);
 
-    int pos = barWidth * progress;
+    int pos = (int)(barWidth * progress);
     std::cout << " [ ";
     for (int l = 0; l < barWidth; ++l) {
       if (l <= pos)
