+++ nntrainer/layers/pooling2d_layer.cpp
@@ -217,9 +217,9 @@ void Pooling2DLayer::calcDerivative(RunLayerContext &context) {
     for (unsigned int b = 0; b < batch; ++b) {
       for (unsigned int i = 0; i < channel; ++i) {
         J = 0;
-        for (int j = -pt; j <= height_stride_end; j += stride[0]) {
+        for (int j = -(int)pt; j <= height_stride_end; j += stride[0]) {
           K = 0;
-          for (int k = -pl; k <= width_stride_end; k += stride[1]) {
+          for (int k = -(int)pl; k <= width_stride_end; k += stride[1]) {
             T del = deriv.getValue<T>(b, i, J, K) / *iter;
             int patch_height_end =
               std::min(static_cast<int>(j + p_height), height);
@@ -470,8 +470,8 @@ void Pooling2DLayer::pooling2d(Tensor &in, bool training, Tensor &output,
     int width_stride_end = width - patch_width - pl;
     for (unsigned int i = 0; i < channel; ++i) {
       const float *in_data_channel_sliced = in_data + i * map_size;
-      for (int j = -pt; j <= height_stride_end; j += stride[0]) {
-        for (int k = -pl; k <= width_stride_end; k += stride[1]) {
+      for (int j = -(int)pt; j <= height_stride_end; j += stride[0]) {
+        for (int k = -(int)pl; k <= width_stride_end; k += stride[1]) {
           float pool_value = pool_fn_fp32(in_data_channel_sliced, i, j, k);
           *out_data = pool_value;
           out_data++;
